Minor edit to cost.cpp
theTime > yr[i] not theTime >= yr[i]

This is a deviation from the WP19 branch but it means that we collect info about the previous year on the 1st of the next year.

The guidelines work involved doing the same but as big changes occurred on that day this got messy. This is neater.